'auth'

+++ api/atividade/delete.php
@@ -8,6 +8,8 @@
 
     include_once '../../config/Database.php';
     include_once '../../models/atividade.php';
+    include_once '../../models/auth.php';
+
 
     //Instaciando Banco de Dados & Conexão
     $database = new Database();
@@ -15,27 +17,34 @@
 
     //Instanciando Atividade object
     $atividade = new Atividade($db);
+    $auth = new Autenticacao($db);
 
-    //Obtém os dados enviados
-    $data = json_decode(file_get_contents("php://input"));
 
-    //Setando o ID para delete
-    $atividade->id = $data->id;
+    if($auth->verificaToken()){
+        //Obtém os dados enviados
+        $data = json_decode(file_get_contents("php://input"));
 
-    //Deletando Atividade
-    if($atividade->delete()) {
-        echo json_encode(
-            array(
-                'Status' => "200",
-                'Mensagem' => 'Atividade Deletada com sucesso!'
-            )
-        );
-    } else {
-        echo json_encode(
-            array(
-                'Status' => '400',
-                'Mensagem' => 'Algo deu errado Atividade Não foi Deletada!'
-            )
-        );
+        //Setando o ID para delete
+        $atividade->id = $data->id;
+
+        //Deletando Atividade
+        if($atividade->delete()) {
+            //resposta
+            echo json_encode(
+                array(
+                    'Status' => "200",
+                    'Mensagem' => 'Atividade Deletada com sucesso!'
+                )
+            );
+        }
+        else{
+            echo json_encode(
+                array(
+                    'Status' => '400',
+                    'Mensagem' => 'Algo deu errado Atividade Não foi Deletada!'
+                )
+            );
+        }
     }
+
 ?>
