+++ api/tipo_atividade/delete.php
@@ -7,6 +7,8 @@
 
     include_once '../../config/Database.php';
     include_once '../../models/tipoAtividade.php';
+    include_once '../../models/auth.php';
+
 
     //Instaciando Banco de Dados & Conexão
     $database = new Database();
@@ -14,27 +16,33 @@
 
     //Instanciando Atividade object
     $tipoAtividade = new tipoAtividade($db);
+    $auth = new Autenticacao($db);
 
-    //Obtém os dados enviados
-    $data = json_decode(file_get_contents("php://input"));
+    if($auth->verificaToken()){
+        //Obtém os dados enviados
+        $data = json_decode(file_get_contents("php://input"));
 
-    //Setando o ID para Update
-    $tipoAtividade->id = $data->id;
+        //Setando o ID para Update
+        $tipoAtividade->id = $data->id;
 
-    //Deletando Atividade
-    if($tipoAtividade->delete()) {
-        echo json_encode(
-            array(
-                "Status" => "200",
-                'Mensagem' => 'Atividade Deletada com sucesso!'
-            )
-        );
-    } else {
-        echo json_encode(
-            array(
-                "Status" => "400",
-                'Mensagem' => 'Atividade Não foi Deletada!'
-            )
-        );
+        //Deletando Atividade
+        if($tipoAtividade->delete()) {
+            //resposta
+            echo json_encode(
+                array(
+                    "Status" => "200",
+                    'Mensagem' => 'Tipo de Atividade Deletada com sucesso!'
+                )
+            );
+        }
+        else {
+            echo json_encode(
+                array(
+                    "Status" => "400",
+                    'Mensagem' => 'Tipo de Atividade Não foi Deletada!'
+                )
+            );
+        }  
     }
+    
 ?>
